Move pointcloud displays to rviz_default_plugins (#153)
* Move pointcloud displays to rviz_default_plugins

* Move pointcloud tests to rviz_default_plugins

* Adjust namespaces for pointcloud display files

* Fix namespaces and import for pointcloud tests

* Rename transformers folder

* Fix linting errors

* Move pointcloud files between CMakeLists, remove display from display_factory

* Add pointcloud displays to plugins_description.xml

* fix includes

* add missing test dependencies

* uncrustify

* Revert "uncrustify"

This reverts commit 03ed305209f15edf7a4d662d290da4f8191ca927.

* disable uncrustify on long include statements
