+++ rviz_common/src/rviz_common/display_factory.cpp
@@ -43,7 +43,6 @@
 
 #include "./temp/default_plugins/displays/tf_display.hpp"
 #include "./temp/default_plugins/displays/robot_model_display.hpp"
-#include "../rviz_default_plugins/point_cloud_display.hpp"
 
 #endif
 
@@ -66,10 +65,6 @@ static Display * newRobotModelDisplay()
 {
   return new rviz_common::RobotModelDisplay();
 }
-static Display * newPointCloudDisplay()
-{
-  return new rviz_default_plugins::PointCloudDisplay();
-}
 
 #endif
 
@@ -79,7 +74,6 @@ DisplayFactory::DisplayFactory()
   addBuiltInClass("rviz", "Group", "A container for Displays", &newDisplayGroup);
   addBuiltInClass("rviz", "TF", "tf display", &newTFDisplay);
   addBuiltInClass("rviz", "RobotModel", "robot model display", &newRobotModelDisplay);
-  addBuiltInClass("rviz", "PointCloud", "point cloud display", &newPointCloudDisplay);
 }
 
 Display * DisplayFactory::makeRaw(const QString & class_id, QString * error_return)
