+++ rviz_default_plugins/package.xml
@@ -29,6 +29,12 @@
   <depend>rviz_common</depend>
   <depend>rviz_rendering</depend>
 
+  <test_depend>ament_cmake_cppcheck</test_depend>
+  <test_depend>ament_cmake_cpplint</test_depend>
+  <test_depend>ament_cmake_gtest</test_depend>
+  <test_depend>ament_cmake_lint_cmake</test_depend>
+  <test_depend>ament_cmake_uncrustify</test_depend>
+
   <export>
     <build_type>ament_cmake</build_type>
   </export>
