+++ rviz_default_plugins/plugins_description.xml
@@ -12,6 +12,28 @@
     </description>
   </class>
 
+  <class
+          name="rviz/PointCloud"
+          type="rviz_default_plugins::displays::PointCloudDisplay"
+          base_class_type="rviz_common::Display"
+  >
+    <description>
+      The Point Cloud display shows data from a (legacy) sensor_msgs/PointCloud message.
+    </description>
+    <message_type>sensor_msgs/PointCloud</message_type>
+  </class>
+
+  <class
+          name="rviz/PointCloud2"
+          type="rviz_default_plugins::displays::PointCloud2Display"
+          base_class_type="rviz_common::Display"
+  >
+    <description>
+      The Point Cloud2 display shows data from a (recommended) sensor_msgs/PointCloud2 message.
+    </description>
+    <message_type>sensor_msgs/PointCloud2</message_type>
+  </class>
+
   <!-- Tool plugins -->
 
   <class
