+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/point_cloud_helpers.hpp
@@ -27,8 +27,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_HELPERS_HPP_
-#define RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_HELPERS_HPP_
+#ifndef RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_HELPERS_HPP_
+#define RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_HELPERS_HPP_
 
 #include <algorithm>
 #include <string>
@@ -157,4 +157,4 @@ inline void getRainbowColor(float value, Ogre::ColourValue & color)
 
 }  // end namespace rviz_default_plugins
 
-#endif  // RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_HELPERS_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_HELPERS_HPP_
