+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/point_cloud_selection_handler.hpp
@@ -27,8 +27,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_SELECTION_HANDLER_HPP_
-#define RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_SELECTION_HANDLER_HPP_
+#ifndef RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_SELECTION_HANDLER_HPP_
+#define RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_SELECTION_HANDLER_HPP_
 
 #ifndef Q_MOC_RUN  // See: https://bugreports.qt-project.org/browse/QTBUG-22829
 # include <memory>
@@ -107,4 +107,4 @@ private:
 }  // namespace rviz_default_plugins
 
 
-#endif  // RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_SELECTION_HANDLER_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_SELECTION_HANDLER_HPP_
