+++ rviz_default_plugins/src/rviz_default_plugins/view_controllers/orbit/orbit_view_controller.hpp
@@ -28,8 +28,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef RVIZ_DEFAULT_PLUGINS__VIEW_CONTROLLERS__ORBIT_VIEW_CONTROLLER_HPP_
-#define RVIZ_DEFAULT_PLUGINS__VIEW_CONTROLLERS__ORBIT_VIEW_CONTROLLER_HPP_
+#ifndef RVIZ_DEFAULT_PLUGINS__VIEW_CONTROLLERS__ORBIT__ORBIT_VIEW_CONTROLLER_HPP_
+#define RVIZ_DEFAULT_PLUGINS__VIEW_CONTROLLERS__ORBIT__ORBIT_VIEW_CONTROLLER_HPP_
 
 #include <OgreVector3.h>
 
@@ -146,4 +146,4 @@ protected:
 }  // namespace view_controllers
 }  // namespace rviz_default_plugins
 
-#endif  // RVIZ_DEFAULT_PLUGINS__VIEW_CONTROLLERS__ORBIT_VIEW_CONTROLLER_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__VIEW_CONTROLLERS__ORBIT__ORBIT_VIEW_CONTROLLER_HPP_
